+++ iGenomics/BWT/BWT_Matcher/EditDistance.m
@@ -20,7 +20,7 @@
 //    printf("\n\n%s\n\n%.*s",a, rangeB.length,b + rangeB.location);
     BOOL testingOn = NO;
     
-    int lenA = rangeA.length, lenB = rangeB.length+1;//First char (range.location-1) will be a "space"
+    int lenA = (int)rangeA.length, lenB = (int)rangeB.length+1;//First char (range.location-1) will be a "space"
     
     int* editDistanceTable = (int*)calloc(lenA*lenB, sizeof(int));
